Fix hidden private key entry
