+++ level.db/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000044
+MANIFEST-000052
